Necessary game files
Platforms.java is the main file with the game itself, which uses the BallStates enumerated type, the Cloud data type, and uses the Hills class to draw the hills for scenery. Standard libraries (especially StdDraw) are required to run.

I apologize for the somewhat excessive amounts of comments in the Platforms file, I wrote this game for fun, but had to present it and walk through my code for one of my classes.